+++ packages/plugins/network-explorer/src/plugin.svelte
@@ -26,9 +26,9 @@
 	data-plugin-name={jsonPackage.name}
 	data-plugin-version={jsonPackage.version}
 >
-	<MaterialTheme pluginType="editor">
+	<LegacyTheme>
 		<NetworkExplorer doc={doc} editCount={editCount} />
-	</MaterialTheme>
+	</LegacyTheme>
 </main>
 
 
@@ -37,7 +37,7 @@
 import jsonPackage from '../package.json'
 // CORE
 import { initPlugin, initSsdTemplate } from '@oscd-plugins/core-ui-svelte'
-import { MaterialTheme } from '@oscd-plugins/ui'
+import { LegacyTheme } from '@oscd-plugins/ui'
 // TYPES
 import type { Utils } from '@oscd-plugins/core-api/plugin/v1'
 import NetworkExplorer from './network-explorer.svelte';